Refactor server.js
